basic tests didn't work from the command line as unit tests behave differently dealing with long and ints. I changed numbers according to types (0 in 0L) and it did work well

+++ DirectMemory-Cache/src/main/java/org/directmemory/CacheManager.java
@@ -153,7 +153,7 @@ public class CacheManager {
 		return offHeapStore.count();
 	}
 	
-	public int usedMemory() {
+	public long usedMemory() {
 		return ((OffHeapStorage)offHeapStore).usedMemory();
 	}
